campustrade/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
# Para desenvolvimento local (SQLite)
# Para Azure SQL Server
pyodbc
